pyproject.toml
[project]
name = "olps-dashboard"
version = "0.1.0"
description = "Research-grade OLPS backtesting platform with Maxblue cost-aware rebalancing"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "yfinance>=0.2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "ruff>=0.0.282",
    "mypy>=1.4.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend", "scripts"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
